+++ lib/core/extensions.dart
@@ -1,8 +1,10 @@
+import 'package:flutter_gemma/core/message.dart';
+
 const userPrefix = "user";
 const modelPrefix = "model";
 const startTurn = "<start_of_turn>";
 const endTurn = "<end_of_turn>";
 
-extension MessageExtension on String {
-  String transformToChatPrompt() => '$startTurn$userPrefix\n$this$endTurn$startTurn$modelPrefix';
+extension MessageExtension on Message {
+  String transformToChatPrompt() => isUser ? '$startTurn$userPrefix\n$text$endTurn\n$startTurn$modelPrefix\n' : '$text$endTurn\n';
 }