+++ test/flutter_gemma_method_channel_test.dart
@@ -25,6 +25,6 @@ void main() {
   test('getPlatformVersion', () async {
     final model = await platform.createModel(isInstructionTuned: false);
     final session = await model.createSession();
-    expect(await session.getResponse('prompt'), 'response');
+    expect(await session.getResponse(), 'response');
   });
 }